Change RollbackScriptConvention inheritance so that it doesn't get run during install phase.
Fixes: OctopusDeploy/Issues#2135
