+++ source/Calamari/Deployment/Conventions/RollbackScriptConvention.cs
@@ -1,10 +1,11 @@
-﻿using Calamari.Integration.FileSystem;
+﻿using System;
+using Calamari.Integration.FileSystem;
 using Calamari.Integration.Processes;
 using Calamari.Integration.Scripting;
 
 namespace Calamari.Deployment.Conventions
 {
-    public class RollbackScriptConvention : PackagedScriptConvention, IRollbackConvention
+    public class RollbackScriptConvention : PackagedScriptRunner, IRollbackConvention
     {
         public RollbackScriptConvention(string scriptFilePrefix, ICalamariFileSystem fileSystem, IScriptEngine scriptEngine, ICommandLineRunner commandLineRunner) :
             base(scriptFilePrefix, fileSystem, scriptEngine, commandLineRunner)
@@ -18,8 +19,6 @@ namespace Calamari.Deployment.Conventions
 
         public void Cleanup(RunningDeployment deployment)
         {
-            
         }
-
     }
 }
